inference_requirements.txt
certifi==2020.4.5.1
numpy==1.18.3
Pillow==7.1.1
six==1.14.0
torch==1.4.0
torchvision==0.5.0
Werkzeug==1.0.1
flask==1.1.2